refactor: break import cycles between storage and sync modules
