+++ packages/browseros-agent/apps/agent/lib/schedules/scheduleStorage.ts
@@ -1,10 +1,8 @@
 import { storage } from '@wxt-dev/storage'
 import { useEffect, useState } from 'react'
-import { sessionStorage } from '@/lib/auth/sessionStorage'
 import { sendScheduleMessage } from '@/lib/messaging/schedules/scheduleMessages'
 import { createAlarmFromJob } from './createAlarmFromJob'
 import type { ScheduledJob, ScheduledJobRun } from './scheduleTypes'
-import { syncSchedulesToBackend } from './syncSchedulesToBackend'
 
 const getAlarmName = (jobId: string) => `scheduled-job-${jobId}`
 
@@ -159,29 +157,4 @@ export function useScheduledJobRuns() {
   }
 
   return { jobRuns, addJobRun, removeJobRun, editJobRun, cancelJobRun }
-}
-
-export async function syncScheduledJobs(): Promise<void> {
-  const jobs = await scheduledJobStorage.getValue()
-  if (!jobs) return
-
-  const session = await sessionStorage.getValue()
-  const userId = session?.user?.id
-  if (!userId) return
-
-  await syncSchedulesToBackend(jobs, userId)
-}
-
-export function setupScheduledJobsSyncToBackend(): () => void {
-  syncScheduledJobs().catch(() => {})
-
-  const unsubscribe = scheduledJobStorage.watch(async () => {
-    try {
-      await syncScheduledJobs()
-    } catch {
-      // Sync failed silently - will retry on next storage change
-    }
-  })
-
-  return unsubscribe
 }
