+++ packages/browseros-agent/apps/agent/screens/sidepanel/history/ChatHistory.tsx
@@ -4,8 +4,8 @@ import { Loader2 } from 'lucide-react'
 import type { FC } from 'react'
 import { useMemo } from 'react'
 import { useSessionInfo } from '@/lib/auth/sessionStorage'
-import { useConversations } from '@/lib/conversations/conversationStorage'
 import { GetProfileIdByUserIdDocument } from '@/lib/conversations/graphql/uploadConversationDocument'
+import { useConversations } from '@/lib/conversations/useConversations'
 import { getQueryKeyFromDocument } from '@/lib/graphql/getQueryKeyFromDocument'
 import { useChatSessionContext } from '@/modules/chat/chat-session-context'
 import { useGraphqlInfiniteQuery } from '@/modules/graphql/graphql-infinite-query.hooks'
