+++ packages/browseros-agent/apps/agent/screens/sidepanel/history/local/LocalChatHistory.tsx
@@ -1,6 +1,6 @@
 import type { FC } from 'react'
 import { useMemo } from 'react'
-import { useConversations } from '@/lib/conversations/conversationStorage'
+import { useConversations } from '@/lib/conversations/useConversations'
 import { useChatSessionContext } from '@/modules/chat/chat-session-context'
 import { ConversationList } from '../components/ConversationList'
 import type { HistoryConversation } from '../components/types'
