+++ packages/browseros-agent/apps/server/tests/__helpers__/index.ts
@@ -1,8 +1,4 @@
+export { killProcessOnPort } from './kill-port'
 export { cleanupBrowserOS, ensureBrowserOS } from './setup'
-export {
-  asToolResult,
-  html,
-  killProcessOnPort,
-  withMcpServer,
-} from './utils'
+export { asToolResult, html, withMcpServer } from './utils'
 export { cleanupWithBrowser, withBrowser } from './with-browser'
